move ConfigurationException class to internal.exception package
